refactor(ui): move overlay primitives into shared/ui/overlays
