+++ src/app/app-providers.tsx
@@ -4,7 +4,7 @@ import { AuthProvider } from "@/entities/session";
 import { ErrorBoundary } from "@/shared/providers/error-boundary";
 import { ReactQueryProvider } from "@/shared/providers/query-provider";
 import { ThemeProvider } from "@/shared/providers/theme-providers";
-import { Toaster } from "@/shared/ui/toaster";
+import { Toaster } from "@/shared/ui/overlays/toaster";
 import { MatrixRain } from "@/shared/ui/effects";
 import { Header } from "@/widgets/header";
