+++ src/app/brand/components-tab.tsx
@@ -40,7 +40,7 @@ import {
   type SelectOption,
 } from "@/shared/ui";
 import { Sparkline, buildMonthlySeries } from "@/shared/ui/sparkline";
-import ConfirmDeleteModal from "@/shared/ui/confirmation-modal";
+import ConfirmDeleteModal from "@/shared/ui/overlays/confirmation-modal";
 import { addToastSuccess, addToastError } from "@/shared/lib/toasts";
 import { Section, Panel, State } from "./_helpers";
