+++ src/shared/ui/ascii-art.tsx
@@ -3,7 +3,7 @@
 import { useEffect, useRef, useState } from "react";
 import { useInViewOnce } from "@/shared/lib/use-in-view";
 import { prefersReducedMotion } from "@/shared/lib/prefers-reduced-motion";
-import { ConsoleTitleBar } from "./console";
+import { ConsoleTitleBar } from "./overlays/console";
 
 /** Density ramp — sparse → dense glyphs, the ASCII "grey scale". */
 const RAMP = " .:-=+*#%@";
