+++ src/shared/ui/overlays/index.ts
@@ -1,0 +1,4 @@
+export { Menu } from "./menu";
+export type { MenuItem } from "./menu";
+export { Modal, ModalHeader, SubmitButton, useModalTitleId } from "./modal";
+export { Console, ConsoleTitleBar } from "./console";
